[OPENBUS-2472] Assistente não pode aceitar tempo de espera entre tentativas igual a ZERO [OPENBUS-2473] [C#] Assistente não pode aceitar tempo de espera entre tentativas igual a ZERO - Implementação da issue. - Intervalo do assistente agora é em segundos e não mais em milisegundos.
git-svn-id: https://subversion.tecgraf.puc-rio.br/engdist/openbus/sdk/csharp/branches/02_00_00@150689 ae0415b3-e90b-0410-900d-d0be9363c56b

+++ src/assistant/AssistantProperties.cs
@@ -27,13 +27,15 @@ namespace tecgraf.openbus.assistant {
   /// </summary>
   public interface AssistantProperties {
     /// <summary>
-    /// Tempo em milisegundos indicando o tempo mínimo de espera antes de cada nova 
+    /// Tempo em segundos indicando o tempo mínimo de espera antes de cada nova 
     /// tentativa após uma falha na execução de uma tarefa. Por exemplo, depois 
     /// de uma falha na tentativa de um login ou registro de oferta, o 
     /// assistente espera pelo menos o tempo indicado por esse parâmetro antes 
-    /// de realizar uma nova tentativa.
+    /// de realizar uma nova tentativa. O valor mínimo é 0,001. Para valores 
+    /// maiores que o mínimo mas com mais de três casas decimais, é utilizada o
+    /// método Math.Ceiling().
     /// </summary>
-    int Interval { get; set; }
+    float Interval { get; set; }
 
     /// <summary>
     /// O ORB a ser utilizado pelo assistente para realizar suas tarefas. O
